Move footpath item entry
