+++ src/openrct2-ui/windows/Scenery.cpp
@@ -20,9 +20,10 @@
 #include <openrct2/localisation/Localisation.h>
 #include <openrct2/management/Research.h>
 #include <openrct2/network/network.h>
+#include <openrct2/object/BannerObject.h>
+#include <openrct2/object/FootpathItemObject.h>
 #include <openrct2/object/ObjectList.h>
 #include <openrct2/object/WallObject.h>
-#include <openrct2/object/BannerObject.h>
 #include <openrct2/sprites.h>
 #include <openrct2/world/LargeScenery.h>
 #include <openrct2/world/Park.h>
