+++ src/openrct2/paint/tile_element/Paint.Path.cpp
@@ -21,6 +21,7 @@
 #include "../../interface/Viewport.h"
 #include "../../localisation/Formatter.h"
 #include "../../localisation/Localisation.h"
+#include "../../object/FootpathItemObject.h"
 #include "../../object/FootpathObject.h"
 #include "../../object/FootpathRailingsObject.h"
 #include "../../object/FootpathSurfaceObject.h"
