+++ src/openrct2/world/Scenery.h
@@ -117,19 +117,6 @@ struct SmallSceneryEntry
     }
 };
 
-enum class PathBitDrawType : uint8_t;
-
-struct PathBitEntry
-{
-    StringId name;
-    uint32_t image;
-    uint16_t flags;
-    PathBitDrawType draw_type;
-    CursorID tool_id;
-    money16 price;
-    ObjectEntryIndex scenery_tab_id;
-};
-
 #pragma pack(pop)
 
 struct LargeSceneryText
@@ -155,27 +142,6 @@ struct rct_scenery_group_entry
     std::vector<ScenerySelection> SceneryEntries;
     uint8_t priority;
     uint32_t entertainer_costumes;
-};
-
-enum
-{
-    PATH_BIT_FLAG_IS_BIN = 1 << 0,
-    PATH_BIT_FLAG_IS_BENCH = 1 << 1,
-    PATH_BIT_FLAG_BREAKABLE = 1 << 2,
-    PATH_BIT_FLAG_LAMP = 1 << 3,
-    PATH_BIT_FLAG_JUMPING_FOUNTAIN_WATER = 1 << 4,
-    PATH_BIT_FLAG_JUMPING_FOUNTAIN_SNOW = 1 << 5,
-    PATH_BIT_FLAG_DONT_ALLOW_ON_QUEUE = 1 << 6,
-    PATH_BIT_FLAG_DONT_ALLOW_ON_SLOPE = 1 << 7,
-    PATH_BIT_FLAG_IS_QUEUE_SCREEN = 1 << 8
-};
-
-enum class PathBitDrawType : uint8_t
-{
-    Light,
-    Bin,
-    Bench,
-    JumpingFountain,
 };
 
 enum
@@ -236,6 +202,7 @@ void SceneryRemoveGhostToolPlacement();
 
 struct WallSceneryEntry;
 struct BannerSceneryEntry;
+struct PathBitEntry;
 
 WallSceneryEntry* GetWallEntry(ObjectEntryIndex entryIndex);
 BannerSceneryEntry* GetBannerEntry(ObjectEntryIndex entryIndex);
